+++ test/test_various.jl
@@ -209,3 +209,19 @@ end
         end
     end
 end
+
+@testset "Test simulation pretty printing" begin
+    sim = mr.Simulation([mr.dwi(bval=1), mr.dwi(bval=2)], geometry=mr.spheres([1, 2.], repeats=[5, 5, 5]), R1=0.1, MT_fraction=0.3)
+    @test repr(sim, context=:compact => true) == "Simulation(2 sequences, Geometry(2 repeating spheres, ), D=0.0um^2/ms, GlobalProperties(T1=10.0ms, MT_fraction=0.3, ))" 
+    @test repr(sim, context=:compact => false) == "Simulation(Geometry(2 repeating spheres, ), D=0.0um^2/ms, GlobalProperties(T1=10.0ms, MT_fraction=0.3, )):
+2 sequences:
+Sequence (TR=2000.0ms):
+    - InstantRFPulse: t=0.0ms, θ=90.0°, ϕ=-90.0°;
+    - InstantRFPulse: t=40.0ms, θ=180.0°, ϕ=0.0°;
+    - Readout at 80.0ms
+Sequence (TR=2000.0ms):
+    - InstantRFPulse: t=0.0ms, θ=90.0°, ϕ=-90.0°;
+    - InstantRFPulse: t=40.0ms, θ=180.0°, ϕ=0.0°;
+    - Readout at 80.0ms
+" 
+end